requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pillow>=10.0.0
piexif>=1.1.3
python-multipart>=0.0.9
